refactor: update resolvers
